requirements.txt
torch==2.2.0
torchvision==0.17.0
opencv-python==4.5.5.62
scikit-image==0.22.0
pandas==2.2.0
numpy==1.26.3
PyYAML==6.0.1
pillow==10.2.0